+++ Projeto/Includes/hitable.h
@@ -1,0 +1,20 @@
+#ifndef HITABLE_H
+#define HITABLE_H
+
+#include "ray.h"
+#include "color.h"
+#include "../../External/glm/glm.hpp"
+
+struct hit_record {
+    float t;
+    glm::vec3 p;
+    glm::vec3 normal;
+    color cor;
+};
+
+class hitable {
+    public:
+        virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const = 0;
+};
+
+#endif
